Feature/API-13

+++ npm/scripts/test.sh
@@ -1,17 +1,31 @@
 #!/bin/bash
 
 # Ensure env vars are set
-export PGHOST="localhost"
-export PGPORT="5432"
-export PGUSER=""
-export PGPASSWORD=""
-export PGDATABASE="minted-api"
+if [ "$CI" = "" ]
+then
+    export PGHOST="localhost"
+    export PGPORT="5432"
+    export PGUSER=""
+    export PGPASSWORD=""
+    export PGDATABASE="minted-api"
 
-# Remove any prior test database
-dropdb "$PGDATABASE"
+    dropdb "$PGDATABASE"
+    createdb "$PGDATABASE"
+else
+    export PGHOST="localhost"
+    export PGPORT="5432"
+    export PGUSER="postgres"
+    export PGPASSWORD=""
+    export PGDATABASE="postgres"
+fi
 
-# Recreate as empty
-createdb "$PGDATABASE"
+# Logging
+echo "Using postgres environment variables:"
+echo "- PGHOST $PGHOST"
+echo "- PGPORT $PGPORT"
+echo "- PGUSER $PGUSER"
+echo "- PGPASSWORD $PGPASSWORD"
+echo "- PGDATABASE $PGDATABASE"
 
 # Recompile
 npm run compile
